refactor: move GenericMatrixGateway to its own file
